tests/wpt/mozilla/tests/mozilla/microdata/tests.html

<!doctype html>
<meta charset="utf-8">
<html>
<head>
  <script src="../../../../web-platform-tests/resources/testharness.js"></script>
  <script src="../../../../web-platform/testharnessreport.js"></script>
  <title>Duplicate ItemProperty and ItemType test</title>
</head>


<h1>Header H1</h1>
<h2>Header H2</h2>
<p>Paragraph</p>
<input>Input</input>
<form>Form</form>
<li>List Item</li>
<ul>Unordered List</ul>
<meta>Meta Tag</meta>
<table>Table</table>
<body id="dup_mainBody" itemtype="http://www.google.com">
<div id="reg" itemtype="hello world">Hi</div>
<div id="dup" itemprop="hello world world">Hi</div>
<div id="single" itemtype="hello">Hi</div>
<div id="extraSpace" itemtype="hello   world">Hi</div>
<div id="caseSensitive" itemtype="Hello hello" </div>
<div id="dupNum" itemtype="34562 34562"></div>
<div id="blank" itemtype=""></div>
<a id="unicode" itemtype=""></a>

</body>
<script>
  test(function(){
    assert_array_equals(document.getElementById("dup").propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
  });
</script>
</html>
